refactor admin to auth folder
