+++ bogan/auth/templates/admin_database.html
@@ -7,7 +7,7 @@
     <nav class="breadcrumb admin-breadcrumb" aria-label="breadcrumbs">
         <ul>
             <li><a href="{{ url_for('main.index') }}">Home</a></li>
-            <li><a href="{{ url_for('admin.admin_dashboard') }}">Admin</a></li>
+            <li><a href="{{ url_for('auth.admin_dashboard') }}">Admin</a></li>
             <li class="is-active"><a href="#" aria-current="page">Datenbank</a></li>
         </ul>
     </nav>
