+++ bogan/auth/templates/admin_users.html
@@ -7,7 +7,7 @@
     <nav class="breadcrumb admin-breadcrumb" aria-label="breadcrumbs">
         <ul>
             <li><a href="{{ url_for('main.index') }}">Home</a></li>
-            <li><a href="{{ url_for('admin.admin_dashboard') }}">Admin</a></li>
+            <li><a href="{{ url_for('auth.admin_dashboard') }}">Admin</a></li>
             <li class="is-active"><a href="#" aria-current="page">Benutzer</a></li>
         </ul>
     </nav>
@@ -81,7 +81,7 @@
                                 {% if user != current_user %}
                                     <div class="buttons are-small">
                                         {% if user.role == 'user' %}
-                                            <form action="{{ url_for('admin.change_user_role', user_id=user.id) }}" method="POST" style="display: inline;">
+                                            <form action="{{ url_for('auth.change_user_role', user_id=user.id) }}" method="POST" style="display: inline;">
                                                 <input type="hidden" name="role" value="admin">
                                                 <button class="button admin-btn-warning is-small" type="submit" onclick="return confirm('Benutzer {{ user.name }} zum Admin machen?')">
                                                     <span class="icon">
@@ -91,7 +91,7 @@
                                                 </button>
                                             </form>
                                         {% else %}
-                                            <form action="{{ url_for('admin.change_user_role', user_id=user.id) }}" method="POST" style="display: inline;">
+                                            <form action="{{ url_for('auth.change_user_role', user_id=user.id) }}" method="POST" style="display: inline;">
                                                 <input type="hidden" name="role" value="user">
                                                 <button class="button admin-btn-primary is-small" type="submit" onclick="return confirm('Admin-Rechte von {{ user.name }} entfernen?')">
                                                     <span class="icon">
