+++ bogan/main/templates/base.html
@@ -60,7 +60,7 @@
                 {% if current_user.is_authenticated %}
                 <a href="{{ url_for('main.profile') }}" class="navbar-item">{{ current_user.name }}'s Profil</a>
                 {% if current_user.role == 'admin' %}
-                <a href="{{ url_for('admin.admin_dashboard') }}" class="navbar-item has-text-danger">
+                <a href="{{ url_for('auth.admin_dashboard') }}" class="navbar-item has-text-danger">
                     <span class="icon">
                         <i class="fas fa-cogs"></i>
                     </span>
